obj: change action count to size_t
This patch is preparation of lifting the action limit.

+++ src/include/libpmemobj/action_base.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017, Intel Corporation
+ * Copyright 2017-2018, Intel Corporation
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -82,10 +82,11 @@ PMEMoid pmemobj_xreserve(PMEMobjpool *pop, struct pobj_action *act,
 void pmemobj_set_value(PMEMobjpool *pop, struct pobj_action *act,
 	uint64_t *ptr, uint64_t value);
 
-void pmemobj_publish(PMEMobjpool *pop, struct pobj_action *actv, int actvcnt);
-int pmemobj_tx_publish(struct pobj_action *actv, int actvcnt);
+void pmemobj_publish(PMEMobjpool *pop, struct pobj_action *actv,
+	size_t actvcnt);
+int pmemobj_tx_publish(struct pobj_action *actv, size_t actvcnt);
 
-void pmemobj_cancel(PMEMobjpool *pop, struct pobj_action *actv, int actvcnt);
+void pmemobj_cancel(PMEMobjpool *pop, struct pobj_action *actv, size_t actvcnt);
 
 #ifdef __cplusplus
 }
